Bug fixes
In this update, I've fixed a couple of bugs and visual glitches and reduced the risk of new possible bugs. Other things got better too :)

If something wasn't working properly before, this update likely fixes it. If not = you can always open an issue and I'll check it out and fix the bug (if there is).

Also now if your SSH runs on, let's say, port 2222, you won't need to change anything to make it work. It will work without any additional configuration because now the script looks up for your SSH server port to allow it automatically. If, let's say, something goes wrong, the script simply chooses the port '22'. (this thing was hard-coded to 22 before; it got smarter now with an autodetection)

And aside of this.. I've removed one old function that wasn't useful anymore and was there doing nothing.